# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM La Bercianita API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: minScore
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortDir
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsPage"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /leads/{id}/suppress:
    post:
      operationId: suppressLead
      tags: [leads]
      summary: Mark lead as no-contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suppressed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /email-drafts:
    get:
      operationId: getEmailDrafts
      tags: [emailDrafts]
      summary: List email drafts
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailDraft"
    post:
      operationId: createEmailDraft
      tags: [emailDrafts]
      summary: Create an email draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDraftInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
  /email-drafts/{id}:
    patch:
      operationId: updateEmailDraft
      tags: [emailDrafts]
      summary: Update an email draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDraftUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
    delete:
      operationId: deleteEmailDraft
      tags: [emailDrafts]
      summary: Delete an email draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /outreach-events:
    get:
      operationId: getOutreachEvents
      tags: [outreachEvents]
      summary: List outreach events
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutreachEvent"
    post:
      operationId: createOutreachEvent
      tags: [outreachEvents]
      summary: Create an outreach event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachEvent"
  /outreach-events/{id}:
    patch:
      operationId: updateOutreachEvent
      tags: [outreachEvents]
      summary: Update an outreach event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachEvent"
  /replies:
    get:
      operationId: getReplies
      tags: [replies]
      summary: List replies
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reply"
    post:
      operationId: createReply
      tags: [replies]
      summary: Create a reply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reply"
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /suppression:
    get:
      operationId: getSuppression
      tags: [suppression]
      summary: List suppression entries
      responses:
        "200":
          description: List of suppressed emails/domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuppressionEntry"
    post:
      operationId: createSuppression
      tags: [suppression]
      summary: Add to suppression list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuppressionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuppressionEntry"
  /suppression/{id}:
    delete:
      operationId: deleteSuppression
      tags: [suppression]
      summary: Remove from suppression list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Lead:
      type: object
      required: [id, businessName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        businessName:
          type: string
        businessType:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        sourceNotes:
          type: ["string", "null"]
        fitScore:
          type: ["integer", "null"]
        fitReason:
          type: ["string", "null"]
        consentStatus:
          type: ["string", "null"]
        crmStatus:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        lastContactedAt:
          type: ["string", "null"]
        nextActionAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadDetail:
      allOf:
        - $ref: "#/components/schemas/Lead"
        - type: object
          properties:
            contacts:
              type: array
              items:
                $ref: "#/components/schemas/Contact"
            emailDrafts:
              type: array
              items:
                $ref: "#/components/schemas/EmailDraft"
            outreachEvents:
              type: array
              items:
                $ref: "#/components/schemas/OutreachEvent"
            replies:
              type: array
              items:
                $ref: "#/components/schemas/Reply"
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
    LeadsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    LeadInput:
      type: object
      required: [businessName]
      properties:
        businessName:
          type: string
        businessType:
          type: string
        city:
          type: string
        province:
          type: string
        country:
          type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        instagram:
          type: string
        sourceUrl:
          type: string
        sourceNotes:
          type: string
        fitScore:
          type: integer
        fitReason:
          type: string
        consentStatus:
          type: string
        crmStatus:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
        lastContactedAt:
          type: string
        nextActionAt:
          type: string
    LeadUpdate:
      type: object
      properties:
        businessName:
          type: string
        businessType:
          type: string
        city:
          type: string
        province:
          type: string
        country:
          type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        instagram:
          type: string
        sourceUrl:
          type: string
        sourceNotes:
          type: string
        fitScore:
          type: integer
        fitReason:
          type: string
        consentStatus:
          type: string
        crmStatus:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
        lastContactedAt:
          type: string
        nextActionAt:
          type: string
    Contact:
      type: object
      required: [id, leadId, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        name:
          type: string
        role:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    EmailDraft:
      type: object
      required: [id, leadId, subject, body, createdAt, updatedAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        subject:
          type: string
        body:
          type: string
        status:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EmailDraftInput:
      type: object
      required: [leadId, subject, body]
      properties:
        leadId:
          type: integer
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        createdBy:
          type: string
    EmailDraftUpdate:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        status:
          type: string
    OutreachEvent:
      type: object
      required: [id, leadId, channel, direction, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        channel:
          type: string
        direction:
          type: string
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        eventAt:
          type: ["string", "null"]
        createdAt:
          type: string
    OutreachEventInput:
      type: object
      required: [leadId, channel, direction]
      properties:
        leadId:
          type: integer
        channel:
          type: string
        direction:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        eventAt:
          type: string
    OutreachEventUpdate:
      type: object
      properties:
        status:
          type: string
        subject:
          type: string
        body:
          type: string
    Reply:
      type: object
      required: [id, leadId, body, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        fromEmail:
          type: ["string", "null"]
        body:
          type: string
        intent:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
        receivedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ReplyInput:
      type: object
      required: [leadId, body]
      properties:
        leadId:
          type: integer
        fromEmail:
          type: string
        body:
          type: string
        intent:
          type: string
        sentiment:
          type: string
        receivedAt:
          type: string
    Task:
      type: object
      required: [id, leadId, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        taskType:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required: [leadId, title]
      properties:
        leadId:
          type: integer
        title:
          type: string
        description:
          type: string
        taskType:
          type: string
        dueDate:
          type: string
        status:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        taskType:
          type: string
        dueDate:
          type: string
        status:
          type: string
    SuppressionEntry:
      type: object
      required: [id, createdAt]
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
    SuppressionInput:
      type: object
      properties:
        email:
          type: string
        domain:
          type: string
        reason:
          type: string
...