# lib/api-zod/src/generated/types/leadDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM La Bercianita API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from "./contact";
import type { EmailDraft } from "./emailDraft";
import type { Lead } from "./lead";
import type { OutreachEvent } from "./outreachEvent";
import type { Reply } from "./reply";
import type { Task } from "./task";

export type LeadDetail = Lead & {
  contacts?: Contact[];
  emailDrafts?: EmailDraft[];
  outreachEvents?: OutreachEvent[];
  replies?: Reply[];
  tasks?: Task[];
};